# docs/swagger.yaml
definitions:
  main.Item:
    properties:
      arr_owner_address:
        additionalProperties:
          type: integer
        type: object
      arr_owner_uid:
        additionalProperties:
          type: integer
        type: object
      category_type:
        type: integer
      chain:
        type: string
      collection_id:
        type: integer
      contract_address:
        type: string
      create_time:
        type: integer
      creator_address:
        type: string
      creator_uid:
        type: integer
      currency_address:
        properties:
          address:
            type: string
          chain:
            type: string
          decimal:
            type: integer
          img:
            type: string
          name:
            type: string
          status:
            type: boolean
          symbol:
            type: string
        type: object
      edition:
        type: integer
      end_at:
        type: integer
      event_ids:
        items:
          type: integer
        type: array
      extend_data:
        properties:
          category:
            type: string
          collection:
            type: string
          create:
            type: string
          like:
            type: string
          owner:
            type: string
          total_offer:
            type: string
        type: object
      external_link:
        type: string
      hidden:
        type: boolean
      id:
        type: integer
      image_url:
        type: string
      is_loot_box:
        type: boolean
      is_sensitive:
        type: boolean
      media_type:
        type: integer
      meta_data:
        type: string
      name:
        type: string
      owner_address:
        type: string
      owner_uid:
        type: integer
      price:
        type: string
      product_no:
        type: string
      start_at:
        type: integer
      status:
        type: integer
      token_id:
        type: string
      total_claim:
        type: integer
      total_edition:
        type: integer
      total_like:
        type: integer
      total_limit:
        type: integer
      total_view:
        type: integer
      type_nft:
        type: integer
      unlockable_content:
        type: string
      update_time:
        type: integer
      uri:
        type: string
      url:
        type: string
    type: object
  main.User:
    properties:
      app:
        type: string
      avatar:
        type: string
      background:
        type: string
      bio:
        type: string
      block:
        type: boolean
      blockchain_address:
        type: string
      can_mint:
        type: boolean
      country_code:
        type: string
      create_time:
        type: integer
      deleted:
        type: boolean
      email:
        type: string
      extend_data:
        additionalProperties:
          type: integer
        type: object
      facebook:
        type: string
      instagram:
        type: string
      is_partner:
        type: boolean
      language:
        type: string
      name:
        type: string
      pubkey:
        type: string
      total_followers:
        type: integer
      total_following:
        type: integer
      twitter:
        type: string
      type_user:
        type: integer
      uid:
        type: integer
      update_time:
        type: integer
      username:
        type: string
      verify:
        type: boolean
      vip:
        type: integer
      youtube:
        type: string
    type: object
info:
  contact: {}
paths:
  /item:
    get:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: position
        in: query
        name: pos
        type: integer
      - description: count
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
      summary: ping example
      tags:
      - Item
    post:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: 'Thông tin item mới '
        in: body
        name: NewItem
        required: true
        schema:
          $ref: '#/definitions/main.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
      summary: ping example
      tags:
      - Item
  /item/{id}:
    delete:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: Id cần xóa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
      summary: ping example
      tags:
      - Item
    put:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: ID cần sửa
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cần sửa
        in: body
        name: ItemEdit
        required: true
        schema:
          $ref: '#/definitions/main.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
      summary: ping example
      tags:
      - Item
  /item/{uid}:
    get:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: 'ID của người dùng cần tìm kiếm '
        in: path
        name: uid
        required: true
        type: integer
      - description: position
        in: query
        name: pos
        type: integer
      - description: count
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Item'
      summary: ping example
      tags:
      - Item
  /user:
    post:
      consumes:
      - application/json
      description: do ping
      parameters:
      - description: 'Thông tin user mới '
        in: body
        name: NewUser
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
      summary: ping example
      tags:
      - User
swagger: "2.0"